Maxito/requirements.txt
matplotlib==3.10.3
matplotlib-inline==0.1.7
numpy==2.2.6
pandas==2.2.3
pandocfilters==1.5.1
groq==0.28.0
python-telegram-bot==22.1